mv: rename 'datatools' module to 'fsutils'
