--- src/onemod/datatools/__init__.py
@@ -1,4 +1,0 @@
-from onemod.datatools.config_interface import ConfigInterface
-from onemod.datatools.data_interface import DataInterface
-
-__all__ = ["ConfigInterface", "DataInterface"]
