+++ src/onemod/fsutils/__init__.py
@@ -1,0 +1,4 @@
+from onemod.fsutils.config_interface import ConfigInterface
+from onemod.fsutils.data_interface import DataInterface
+
+__all__ = ["ConfigInterface", "DataInterface"]
